# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-symmetricdifference",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — the symmetric difference of the lists `a` and `b`.",
        "description": "Reads each element of both lists as a JSON integer (`i64`). If any element is\nnot an integer the request is rejected with `422`. Otherwise the sorted list\nof distinct values appearing in exactly one of `a` or `b` is returned as\n`result`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymmetricDifferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "an element is not a valid integer"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "properties": {
          "a": {
            "type": "object",
            "description": "The first set, as a list of integers. Every element must be a JSON\ninteger (i64). Duplicates are allowed and treated as a set."
          },
          "b": {
            "type": "object",
            "description": "The second set, as a list of integers. Every element must be a JSON\ninteger (i64). Duplicates are allowed and treated as a set."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "SymmetricDifferenceResponse": {
        "type": "object",
        "required": [
          "a",
          "b",
          "result"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          },
          "result": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      }
    }
  }
}